Refactored request to utility module
